Add true Cities component
